Add files via upload
This includes pretty much all of the relevant files to both see what I did and to actually recreate my whole project. There is a presentation link to a short video about this project in the read me. I am writing a lengthy blog post explaining all this. The only changes might add are comments or more relevant files. As far as I am concerned, this code is done and I'm moving on to the next project.