Add files via upload
this works by first spiking the wrong answers for a while and then spiking the right answer much longer for the duration of the simulation. This gives the behavior we want in that there are more spikes on the right answer than the wrong and wrong spikes usually happen before all the internal activity so the STDP should reduce their weights. That's the theory anyway, in practice the trained network is just very marginally better than random guessing.