Add files via upload
Not exactly what I wanted but this trained with STDP to classify pretty well. It has distinct bands of average spiking intensity for each of the 3 classes. Was trying to get it to be certain neurons active for certain classes but this is still good progress, just kind of cheesy.